Tests

+++ app/src/main/java/irina/menuplanner/model/ShoppingListProductItem.kt
@@ -2,8 +2,8 @@ package irina.menuplanner.model
 
 
 data class ShoppingListProductItem(
-    val shoppingListItemId: Int,
-    val productId: Int,
+    val shoppingListItemId: Long,
+    val productId: Long,
     var isChecked: Boolean,
     var name: String
 )
